requirements.txt
streamlit
pandas
openpyxl
psycopg2-binary
duckdb
requests
BeautifulSoup4
python-dotenv